ingestion/requirements.txt
# Python dependencies for Chișinău Routing Engine - Ingestion Layer

# OSM data extraction
osmnx>=1.9.0
networkx>=3.2

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0
geoalchemy2>=0.14

# GTFS-RT parsing
gtfs-realtime-bindings>=1.0.0
protobuf>=4.25.0

# HTTP client
requests>=2.31.0
httpx>=0.27.0

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# Geo utilities
shapely>=2.0.0
pyproj>=3.6.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging
structlog>=24.1.0

# Async support
asyncio-throttle>=1.0.0
